MAINT: move _scalar_types to the _detail namespace
