+++ torch_np/__init__.py
@@ -1,5 +1,5 @@
 from ._dtypes import *
-from ._scalar_types import *
+from ._detail._scalar_types import *
 from ._wrapper import *
 #from . import testing
